requirements.txt
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.3.0
defusedxml==0.7.1
filelock==3.19.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Markdown==3.7
MarkupSafe==3.0.3
numpy
packaging==24.2
pandas
pillow==11.3.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
pymupdf4llm==0.0.17
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.27.1
rsa==4.9
six==1.17.0
soupsieve==2.6
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
youtube-transcript-api==1.2.2
yt-dlp==2025.9.26
ruff